Add files via upload
This version of blockerboy adds in a bit more Realtime inspection and incorporates a trained IsolationForest AI model to perform Anomaly Detection while still performing the original NIDS functions However will still need to be coupled with the Original BlockerBoy.py to enable realtime blocking. The NIDS as of now is still basic and performing as a method of Realtime logging.